pyproject.toml
[project]
name = "powa"
version = "0.0.1"
description = "POwer WArden Daemon"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
authors = [ { name = "Julian Nicolas Rodriguez", email = "junirodriguezz1@gmail.com" } ]
keywords = ["powa", "daemon", "power management", "energy", "monitoring"]
dependencies = [
    "pyyaml>=6.0",
    "smbus2==0.5.0",
    "aiohttp==3.12.13",
]

[project.urls]
repository = "https://github.com/marifante/powa"

[project.scripts]
powa = "powa.cli:main"

[tools.setuptools.packages.find]
include = ["powa*"]

[tool.poetry.dev-dependencies]
tox = "^4.0"
pytest = "^8.0"
pytest-asyncio = "^0.23"
pytest-cov = "^5.0"

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py311

[testenv]
deps =
    pytest
    pytest-asyncio
    pytest-cov
commands =
    pytest --cov=powa --cov-report=term-missing
"""
